separate TopFragment as one of feature module
