+++ .idea/gradle.xml
@@ -12,7 +12,11 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/buildSrc" />
+            <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/data" />
+            <option value="$PROJECT_DIR$/feature" />
+            <option value="$PROJECT_DIR$/feature/feature_common" />
+            <option value="$PROJECT_DIR$/feature/settings_top" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
