--- app/src/main/java/black/bracken/picsorter/ext/Context.kt
@@ -1,7 +1,0 @@
-package black.bracken.picsorter.ext
-
-import android.app.NotificationManager
-import android.content.Context
-
-val Context.notificationManager
-    get() = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
