+++ app/src/main/java/black/bracken/picsorter/ui/manipulating/ManipulatingActivity.kt
@@ -6,8 +6,8 @@ import androidx.databinding.DataBindingUtil
 import black.bracken.picsorter.R
 import black.bracken.picsorter.data.model.SimpleManipulating
 import black.bracken.picsorter.databinding.ActivityManipulatingBinding
-import black.bracken.picsorter.ext.observeOnce
-import black.bracken.picsorter.ext.setOnTextChanged
+import black.bracken.picsorter.feature_common.ext.observeOnce
+import black.bracken.picsorter.feature_common.ext.setOnTextChanged
 import coil.api.load
 import com.afollestad.materialdialogs.MaterialDialog
 import com.afollestad.materialdialogs.files.folderChooser
