--- app/src/main/res/navigation/nav_graph.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:id="@+id/nav_graph"
-        app:startDestination="@id/topFragment">
-    <fragment
-            android:id="@+id/topFragment"
-            android:name="black.bracken.picsorter.ui.settings.top.TopFragment"
-            android:label="TopFragment">
-        <action
-                android:id="@+id/action_topFragment_to_directoriesChooserFragment"
-                app:destination="@id/directoriesChooserFragment" />
-        <action
-                android:id="@+id/action_topFragment_to_simpleManipulatingSettingsFragment"
-                app:destination="@id/simpleManipulatingSettingsFragment" />
-    </fragment>
-    <fragment
-            android:id="@+id/directoriesChooserFragment"
-            android:name="black.bracken.picsorter.ui.settings.dirchooser.DirectoriesChooserFragment"
-            android:label="DirectoriesChooserFragment" />
-    <fragment
-            android:id="@+id/simpleManipulatingSettingsFragment"
-            android:name="black.bracken.picsorter.ui.settings.simplemanipulating.top.SimpleManipulatingTopFragment"
-            android:label="SimpleManipulatingSettingsFragment">
-        <action
-                android:id="@+id/action_simpleManipulatingSettingsFragment_to_simpleManipulatingRegistererFragment"
-                app:destination="@id/simpleManipulatingRegistererFragment" />
-        <action
-                android:id="@+id/action_simpleManipulatingSettingsFragment_to_topFragment"
-                app:destination="@id/topFragment" />
-    </fragment>
-    <fragment
-            android:id="@+id/simpleManipulatingRegistererFragment"
-            android:name="black.bracken.picsorter.ui.settings.simplemanipulating.registerer.SimpleManipulatingRegistererFragment"
-            android:label="SimpleManipulatingRegistererFragment">
-        <action
-                android:id="@+id/action_simpleManipulatingRegistererFragment_to_simpleManipulatingSettingsFragment"
-                app:popUpTo="@id/simpleManipulatingSettingsFragment" />
-    </fragment>
-</navigation>
