+++ app/src/main/res/values/strings.xml
@@ -31,8 +31,6 @@
     <string name="title_copyright">Copyright</string>
     <string name="layout_settings_open_notification_settings">Open a standard notification settings shows on taking screenshot</string>
     <string name="layout_settings_description_open_notification_settings">Advanced: To disable Android standard notification.</string>
-    <string name="dialog_permission_request_title">"Allow access to photos. "</string>
-    <string name="dialog_permission_request_subtitle">This app doesn\'t work without permissions. Push a below button to allow.</string>
     <string name="layout_simplemanipulatingsettings_title">List of simple manipulation</string>
     <string name="layout_settings_open_simple_manipulating_settings">Open a settings for simple manipulating</string>
     <string name="layout_settings_description_open_simple_manipulating_settings">To omit manipulations by registering them.</string>
