+++ buildSrc/src/main/java/dependencies/Versions.kt
@@ -1,5 +1,6 @@
 package dependencies
 
+@Suppress("unused")
 object Versions {
     const val androidCompileSdkVersion = 28
     const val androidMinSdkVersion = 28
