+++ data/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="black.bracken.picsorter.data"></manifest>
+<manifest package="black.bracken.picsorter.data" />
