+++ feature/feature_common/build.gradle.kts
@@ -1,0 +1,17 @@
+import dependencies.Dep
+
+plugins {
+    id("com.android.library")
+    kotlin("android")
+    kotlin("android.extensions")
+    kotlin("kapt")
+}
+
+apply(from = rootProject.file("gradle/android.gradle"))
+
+dependencies {
+    Dep.corePackages.forEach { implementation(it) }
+    Dep.androidCorePackages.forEach { implementation(it) }
+
+    implementation(Dep.MaterialDialogs.core)
+}
