+++ feature/feature_common/src/androidTest/java/black/bracken/picsorter/feature_common/ExampleInstrumentedTest.kt
@@ -1,0 +1,22 @@
+package black.bracken.picsorter.feature_common
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.*
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("black.bracken.picsorter.feature_common", appContext.packageName)
+    }
+}
