+++ feature/feature_common/src/main/java/black/bracken/picsorter/feature_common/ext/Context.kt
@@ -1,0 +1,7 @@
+package black.bracken.picsorter.feature_common.ext
+
+import android.app.NotificationManager
+import android.content.Context
+
+val Context.notificationManager
+    get() = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
