+++ feature/feature_common/src/main/res/drawable/ic_touch_app_black.xml
