+++ feature/feature_common/src/test/java/black/bracken/picsorter/feature_common/ExampleUnitTest.kt
@@ -1,0 +1,16 @@
+package black.bracken.picsorter.feature_common
+
+import org.junit.Assert.*
+import org.junit.Test
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
