+++ feature/settings_top/src/androidTest/java/black/bracken/picsorter/settings_top/ExampleInstrumentedTest.kt
@@ -1,0 +1,22 @@
+package black.bracken.picsorter.settings_top
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.*
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("black.bracken.picsorter.settings_top", appContext.packageName)
+    }
+}
