+++ feature/settings_top/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="black.bracken.picsorter.settings_top" />
