+++ settings.gradle.kts
@@ -1,1 +1,4 @@
+include(":core")
+include(":feature:feature_common")
+include(":feature:settings_top")
 include(":app", ":data")
